Don't try to modify request.args.
Summary:
This works on certain versions of Flask and werkzeug, but not others
(which is why it didn't get caught before T329).  This commit retreats
from that approach and addresses the problem more directly...

reqparse was behaving strangely in the presence of arguments it didn't
know about (like 'callback' and '_').  By declaring those arguments and
then not using them, the test case added in the previous commit passes.

Sorry for the mess on this.. :/

Test Plan:
  The test suite.

Reviewers: tflink, mkrizek

Reviewed By: tflink, mkrizek

Subscribers: tflink

Maniphest Tasks: T329, T345

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D236