Added UUID to Job
A piece of code needed for the ExecDB. The Job now can store an UUID.
Not required at the moment, but once provided, the jobs can be accessed
by it using the RESTfull interface like:

    http://.../api/v1.0/jobs/UUID

While this is not exactly RESTfull in the spirit (resource should not
be accessible by multiple identifiers), it is IMHO the reasonable thing
to do, since the UUID is Taskotron-related identifier, but at the same
time is unique in our scope.

This is a pre-requirement for the resultsdb_api, and libtaskotron
ExecDB-related patches.