Add parameter which disables returning data from update_job
As discussed in T456 and T452, we need to (as a short term workaround at least)
be able to not return data from update_job, as it makes multiple selects on
the Results table, which take too long. It is now ON by default, so the
update_job() call returns just the status code and empty data.

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D333