Added indexes to Foreign keys
PostgreSQL does not automatically create Indexes for Foreign keys.
This was found out while solving T452. Without indexes, queries like:
    SELECT * FROM result WHERE 115515 = result.job_id;
take about 250 ms. With index on the Foreign key, it is 0.048 ms.

In Postgres, if the database does not use the C locale, special option
needs to be set during the Index creation, in order fo LIKE filter to be
able to use it.

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D338