Allow Testcase autocreation for create_result
In the "old" resultsdb, testcases were identified just by the URL,
with the transition to RESTfull interface, we switched to identification
by name, and compulsory URL attribute (which is meant to point to the
testcase's documentation). We don't actually really use the URL attribute
anyways, and needing to check for Testcase's existence prior to any
`create_result()` call is kind of pointless anyways - since "not found"
testcase is then created anyway, just on the client side.

This patch moves the "automagic" part to the server, while still allowing
the client to instruct the server to require the testcase to be created
beforehand by introducing the `strict` parameter (False by default) to the
`create_result()` routine

This does not fix the issue in T519 per se (patch will follow), per se,
but changes the way testcase-creation is handled in clients.

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D431