With all JSON queries, return more pagination metadata.
Summary:
This now includes:

- total - the total number of entities that can be retrieved.
- pages - the total number of pages that can be retrieved

Without these, if you have code that is paging through resultsdb
results, there is no way to know when you are going to be done -- and no
way to jump to the last page and iterate through them in reverse.

Test Plan: Run the test suite.

Reviewers: jskladan, tflink

Reviewed By: tflink

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D263