Change testcase.name to text
First step to testcase namespacing - changing name from varchar(255) to text.
The alembic migration has one slight problem - sqlite does not support ALTER COLUMN, and so the upgrade fails with SQLite. Works with postgres though (tested).
Are we OK with it failing on SQLITE (maybe with documentation saying "if you are on sqlite, please remove the database and initialize it again", or should I try to add some workaround?
I do not think it's worth it (the time spent on working around sqlite), but I'm open to ideas.

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D750