Flexible messaging.
Summary:
This replaces the fedmsg message publication with a plugin-based approach.

Plugins can be specified using setuptools entry-points and must extend an
abstract base class defined in `resultsdb.messaging`.

The original fedmsg publication code is reimplemented as a new plugin and a new
`DummyPlugin` is added for use in the test suite.  The original test cases all
still pass. Two new test cases are added that use and verify the DummyPlugin.

Test Plan: Run the functional test suite.

Reviewers: tflink, jskladan

Reviewed By: jskladan

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D1061