Fix functional tests in the test suite.
Summary:
The issue is that the messaging plugins are defined as setuptools
entry-points, and those live in the .egg-info directory.  If you're developing
locally, that directory is usually present in `pwd` due to `python setup.py
develop` or `python setup.py egg-info`.  But in the spec file, we install to a
directory that is not in the PYTHONPATH.  This change re-enables the `-F`
option but also specifies the PYTHONPATH so that the entry-points are found and
the tests pass.

This should fix T874.

Test Plan: Try building the rpm in mock.  It works!

Reviewers: tflink, jskladan, mkrizek

Reviewed By: mkrizek

Maniphest Tasks: T874

Differential Revision: https://phab.qa.fedoraproject.org/D1073