Make messaging plugin loading more pleasant for dev
The way the messaging plugins are loaded, it is required to run
`python setup.py develop` for resultsdb to function properly,
if you use it from git checkout.

To make the development a bit more pleasant, I made sure the
dummy plugin can be loaded even without that.

Test Plan:
As @ralph noted at IRC, doing this kind of lets us in the dark most
of the time, so to catch the plugin loading issues, I updated the
unitests with code that tries to load both dummy (should work always)
and fedmsg (only works after `setup.py develop`) plugins.
Helpfull message is shown when loading the fedmsg plugin fails.

Fixes T883
Differential Revision: https://phab.qa.fedoraproject.org/D1087