Use gunicorn to run app
This will switch using gunicorn instead of flask.run() for running the app.

It also swtiches storing settings.py in a Secret instead of a ConfigMap
in order to keep secret information under control.

Template asks for image name instead of version, this way resultsdb image
stored in an external registry can be also used.

Note that this is not a production ready solution, as resultsdb
requires authorization to be handled by the server. In current deployements
this is done using httpd configuration, but I was not able to find
a similar solution with gunicorn.