ship Alembic migration scripts as package data
can always find its Alembic migrations regardless whether it was
imported from an installed RPM package, or a random git checkout
anywhere on the filesystem.

Specifically, for Greenwave I would like to import the ResultsDB
application and run it as a fixture in the Greenwave tests.
Right now, that doesn't work because ResultsDB is expecting
`alembic.ini` to be in the current working directory for a git
checkout.
This seemed like a nice solution to how it can find its Alembic
migrations regardless of the cwd.

Differential Revision: https://phab.qa.fedoraproject.org/D1209