Add 'scenario' to identifiers in get_prev_result
As @ralph pointed out, the get_prev_result logic should also
use 'scenario' as one of the identifiers when deciding if a
result is 'the same as' a previous one; this is explicitly what
the key is intended for, and it is necessary for openQA tests,
as they may have the same testcase name, arch, item and type,
but be for a different firmware type, for instance. This may be
preventing us from emitting fedmsgs for some openQA results.

Signed-off-by: Adam Williamson <awilliam@redhat.com>