fix 'job_url' and 'group_url' keys in messages
The 'job_url' key is supposed to appear in messages when a result is
created using API v1, but this regressed in commit
62c49f7fac88d88c954dccbb2a532e9f67b82faf which introduced API v2. The
create_fedmsg() function was still referring to the `result.job`
attribute even though it had been removed from the model in that commit.
This patch restores the previous intended behaviour for API v1 messages,
with a test case.

The 'group_url' key was introduced in the same commit as an analog when
results are created using API v2, but this also never worked because of
a typo: it was referring to a `result.group` attribute but the model
actually has `result.groups` (plural). The same typo was carried over in
commit fd4ebfb0e8389931213acc9eefc40d07e158d30a ("flexible messaging").
Since this never worked, and thus the 'group_url' key has never been
included in any messages, this patch simply drops it.