generic message format which matches the HTTP API v2
Fixes #92.

This changes the message structure to contain the complete result data,
in the same format as the HTTP API v2.

Existing message structure and de-duplication logic is preserved in the
new publish_taskotron_message() function, which is disabled by default
but can be enabled for backwards compatibility by setting
MESSAGE_BUS_PUBLISH_TASKOTRON=True. This function is hardcoded to only
send on fedmsg and with 'taskotron' as the topic.