Update OpenShift test template to allow deploying with other apps
Build image used in OpenShift from Fedora

The .spec file change to update dependencies in order to
use python2-* instead of python-* broken building on top of
centos.

Use 'app' label in the OpenShift test template

When deployed together with other services (WaiverDB,
Greenwave), this label can be used for selecting
ResultsDB related resources.

Allow anyone to POST a result

For testing purposes it's easier to allow anyone to
POST on the external facing service.