Latest endpoint: group by additional fields
The results/latest endpoint groups the results by testcase. Sometimes
it might be useful to group by additional fields in the ResultData
table. This new version of the API endpoint gets an additional
parameter "group_by" that takes a list of fields (separated by ",")
on which will be done the group by.
This new feature is not available in case no other filtering is
provided. In that case the API returns an 400 error, since it wouldn't
be possible to retrieve the requested results with an efficient request.