Handle more fedora-messaging exceptions when publishing messages
As well as PublishReturned and ConnectionException, messaging
can raise PublishTimeout (publishing timed out) or
PublishForbidden (publishing was denied for lack of permission)
when you attempt to publish a message. We should handle these
as well. At present it seems like if this happens, the exception
being unhandled ultimately causes WSGI to return a 500 or 504
error to the client that submitted the result; the client may
then try and submit the result again, leading to it being
duplicated many times. If we handle the exception instead, the
client should get a successful response and not re-submit the
result. The message not being published is still a problem, but
probably not as bad as clients constantly re-submitted results
forever.

Signed-off-by: Adam Williamson <awilliam@redhat.com>