Fix authorization failure responses
Messages and logging need to be improved for the new authenticated API
to troubleshoot issues.

Failure to authorize a request with the service must be logged.

If permission is denied to post a data (or access an endpoint), 403
Forbidden status code must be returned (currently it is 401).

Failure to find user in LDAP should be send back to user (with 403
response including the test case name)

In case of LDAP errors, the service should return 502 status code (not
4XX).

Return JSON formatted responses consistently.

JIRA: RHELWF-10633